# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Netflix API
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  # - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
tags:
  - name: 'Users'
    description: Operations about Netflix Users
  - name: 'Authentication'
    description: Authentication operations
  - name: 'Videos'
    description: Operations about Netflix Videos
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: X-Session-ID
paths:
  /users:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    post:
      description: Creates a new user with a pair of credentials
      # used as the method name of the controller
      operationId: createUser
      tags:
      - "Users"
      parameters:
        - name: user
          in: body
          description: The user to create
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/CreateUserResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Creates a new user with a pair of credentials
      # used as the method name of the controller
      operationId: deleteUser
      tags:
      - "Users"
      parameters:
        - name: user
          in: body
          description: delete
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DeleteUserResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      description: Creates a new user with a pair of credentials
      # used as the method name of the controller
      operationId: getUser
      tags:
      - "Users"
      parameters: []
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
            # a pointer to a definition
              $ref: "#/definitions/GetUserResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /authentication/login:
    x-swagger-router-controller: authentication
    post:
      description: Authenticates a user with a pair of credentials and returns a sessionID
      operationId: login
      tags:
      - "Authentication"
      parameters:
        - name: credentials
          in: body
          description: Credentials to authentication the user with
          required: true
          schema:
            $ref: "#/definitions/Credentials"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/AuthenticationResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /authentication/logout:
    x-swagger-router-controller: authentication
    get:
      description: Authenticates a user with a pair of credentials and returns a sessionID
      operationId: logout
      tags:
      - "Authentication"
      parameters:
        - name: X-SESSION-ID
          in: header
          description: Session ID
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/AuthenticationResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /products:
    x-swagger-router-controller: products
    get:
      description: Find products
      operationId: getProducts
      tags:
      - "Products"
      parameters: []
      responses:
        "200":
          description: Success
          schema:
            type: 'array'
            items:
              # a pointer to a definition
              $ref: "#/definitions/Product"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Create product
      operationId: createProduct
      tags:
      - "Products"
      parameters:
        - name: product
          required: true
          in: body
          description: adding product
          schema:
            $ref: "#/definitions/Product"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/CreateProductResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
  /products/{productName}:
    get:
      description: Get product with the given name
      operationId: findProduct
      tags:
      - "Products"
      parameters:
        - name: productName
          in: path
          description: Product name
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Product"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
  /users/balance:
    x-swagger-router-controller: users
    put:
      tags:
        - balance
      summary: Updated balance
      description: Updates the balance
      operationId: updateBalance
      produces:
        - application/xml
        - application/json
      parameters:
        - name: updateCredentials
          required: true
          in: body
          description: Credentials to update the user with
          schema:
            $ref: "#/definitions/UpdateBalanceCredentials"
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Invalid amount supplied
        '404':
          description: User balance not found
    get:
      tags:
        - balance
      summary: Get balance
      description: ''
      operationId: getBalance
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - name: username
          in: header
          description: The user name for getting balance
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '405':
          description: Invalid input
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  User:
    required:
      - username
      - password
    properties:
      username:
        type: string
        example: 'jwalker'
      password:
        type: string
        example: 'supersecret'
      fullname:
        type: string
        example: 'Johnny Walker'
      balance:
        type: integer
        format: int64
        default: 0
  CreateUserResponse:
    properties:
      _id:
        type: string
        example: '123'
  CreateProductResponse:
    properties:
      _id:
        type: string
        example: '123'
  Credentials:
    required:
      - username
      - password
    properties:
      username:
        type: string
        example: 'jwalker'
      password:
        type: string
        example: 'supersecret'
  AuthenticationResponse:
    required:
      - sessionID
    properties:
      sessionID:
        type: string
        example: 'axd43xs'
  Product:
    required:
      - name
      - cost
    properties:
      name:
        type: string
        example: 'Avengers Infinity War'
      cost:
        type: integer
        format: int64
        example: 1234
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  DeleteUserResponse:
    required:
      - username
    properties:
      username:
        type: string
  GetUserResponse:
    required:
        - username
    properties:
      username:
        type: string
        example: 'jwalker'
  UpdateBalanceCredentials:
    required:
      - username
      - amount
    properties:
      username:
        type: string
        example: 'jwalker'
      amount:
        type: integer
        format: int64
        example: 213